using engine to setup view extensions
