--- lib/jquery_ui_rails_helpers/rails_plugin/engine.rb
@@ -1,7 +1,0 @@
-module JqueryUiHelpers
-  module Rails
-    class Engine < ::Rails::Engine
-    	puts "JqueryUiHelpers engine loaded"
-    end
-  end
-end
